`ptree.Iterator` Improvements (#13)
The `ptree.Iterator` inefficiently seeks through every blob for every entry instead of holding a position in the blob between entries. The `ptree.StreamReade`r is implemented this way. Changes the ptree.Iterator to be implemented in terms of the `ptree.StreamReader`

Produces a ~6x improvement in `got debug fs` which runs in O(n) where n is the number of entries for extents and metadata.