Adds gothost (#57)
`gothost` provides a standard format for using a branch and its content to configure a service hosting a Space.  The branch `__host__` is reserved for this purpose.

The IAM policy file previously stored in `.got/policy` is now stored in a file `POLICY` at the root of the host configuration branch.